pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-mail"
version = "0.1.0"
description = "MCP server for sending and reading email via SMTP/IMAP"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "fastmcp==3.2.0",
    "fastapi==0.135.2",
    "uvicorn[standard]==0.42.0",
    "aiosmtplib==5.1.0",
    "aioimaplib==2.0.1",
    "pydantic-settings==2.13.1",
    "email-validator==2.3.0",
]

[project.scripts]
mcp-mail = "mcp_mail.main:cli"

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "httpx==0.28.1",
    "aiosmtpd==1.4.6",
    "ruff==0.11.12",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
